New approach for posix-on-win32
This uses a hopefully stable and productive approach. We layer
on top of existing Windows headers where possible, and disable
things that cause us problems. This uses an #include_next
emulation that requires build system support (locating your
copies of Windows headers). This can be built into Premake and
probably Cmake. It's a flaw at the moment, definitely, but
somewhat necessary.