Add/update tools for generating headers
Update find_includes.py so it saves the result of the long search,
making future passes on the same filesystem fast.

Create in-windows.py that finds Windows headers (based on a
few patterns as to where Microsoft puts VC and Windows Kit
headers).

Create gen-wrap.py that creates wrapper headers for all C and
POSIX headers.